requirements.txt
ansible
gitpython
PyYAML
colorlog
yamlordereddictloader